Replace Enemy and Enemy Controller by new classes
