+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/Boss.java
@@ -1,9 +1,9 @@
 package com.cursedarchie.platformer.actors.enemies;
 
 import com.badlogic.gdx.math.Vector2;
-import com.cursedarchie.platformer.actors.NewEnemy;
+import com.cursedarchie.platformer.actors.Enemy;
 
-public class Boss extends NewEnemy {
+public class Boss extends Enemy {
 
     public Boss(Vector2 pos) {
         super(pos, 4f, 10f);
