+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/EnemyState.java
@@ -1,9 +1,9 @@
 package com.cursedarchie.platformer.actors.enemies.logic;
 
-import com.cursedarchie.platformer.actors.NewEnemy;
+import com.cursedarchie.platformer.actors.Enemy;
 
 public interface EnemyState {
-    void enter(NewEnemy enemy);
-    void update (NewEnemy enemy, float delta);
-    void exit (NewEnemy enemy);
+    void enter(Enemy enemy);
+    void update (Enemy enemy, float delta);
+    void exit (Enemy enemy);
 }
