+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/EnemyStateMachine.java
@@ -1,13 +1,13 @@
 package com.cursedarchie.platformer.actors.enemies.logic;
 
-import com.cursedarchie.platformer.actors.NewEnemy;
+import com.cursedarchie.platformer.actors.Enemy;
 import com.cursedarchie.platformer.actors.enemies.logic.states.IdleState;
 
 public class EnemyStateMachine {
-    private NewEnemy enemy;
+    private Enemy enemy;
     private EnemyState currentState;
 
-    public EnemyStateMachine (NewEnemy enemy) {
+    public EnemyStateMachine (Enemy enemy) {
         this.enemy = enemy;
         this.currentState = new IdleState();
         this.currentState.enter(enemy);
