+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/states/DeadState.java
@@ -2,17 +2,17 @@ package com.cursedarchie.platformer.actors.enemies.logic.states;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.Vector2;
-import com.cursedarchie.platformer.actors.NewEnemy;
+import com.cursedarchie.platformer.actors.Enemy;
 import com.cursedarchie.platformer.actors.enemies.logic.EnemyState;
 
 public class DeadState implements EnemyState {
     @Override
-    public void enter(NewEnemy enemy) {
+    public void enter(Enemy enemy) {
         Gdx.app.log("Enemy", "Started Dying");
     }
 
     @Override
-    public void update (NewEnemy enemy, float delta) {
+    public void update (Enemy enemy, float delta) {
         enemy.setHealth(0);
         enemy.setDamage(0);
         enemy.setPosition(new Vector2(99, 99));
@@ -20,7 +20,7 @@ public class DeadState implements EnemyState {
     }
 
     @Override
-    public void exit (NewEnemy enemy) {
+    public void exit (Enemy enemy) {
         Gdx.app.log("Enemy", "Finished Dying");
     }
 }
