+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/logic/states/RetreatState.java
@@ -1,20 +1,20 @@
 package com.cursedarchie.platformer.actors.enemies.logic.states;
 
-import com.cursedarchie.platformer.actors.NewEnemy;
+import com.cursedarchie.platformer.actors.Enemy;
 import com.cursedarchie.platformer.actors.enemies.logic.EnemyState;
 
 public class RetreatState implements EnemyState {
     @Override
-    public void enter(NewEnemy enemy) {
+    public void enter(Enemy enemy) {
     }
 
     @Override
-    public void update (NewEnemy enemy, float delta) {
+    public void update (Enemy enemy, float delta) {
 
     }
 
     @Override
-    public void exit (NewEnemy enemy) {
+    public void exit (Enemy enemy) {
 
     }
 }
